Renaming Lights.h to LightDefinitions to make purpose of file more clear.
