+++ Assets/Shaders/Shared/LightDefinitions.h
